reroute concept set/concepts navigation
